Split, refactor & improve tests for SearchEngineConfiguration
