+++ app/models/gobierto_budgets/search_engine_configuration/data.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module GobiertoBudgets
+  module SearchEngineConfiguration
+    class Data
+
+      def self.index
+        "data"
+      end
+
+      def self.type_population
+        "population"
+      end
+
+      def self.type_debt
+        "debt"
+      end
+
+    end
+  end
+end
