+++ config/application.rb
@@ -46,7 +46,7 @@ module Gobierto
       "#{config.root}/lib/constraints",
       "#{config.root}/lib/errors",
       "#{config.root}/lib/ibm_notes",
-      "#{config.root}/lib/liquid",
+      "#{config.root}/lib/liquid"
     ]
     config.autoload_paths += required_paths
     config.eager_load_paths += required_paths
